+++ lib/rails_ai_build/engine.rb
@@ -8,6 +8,10 @@ module RailsAiBuild
       g.test_framework :rspec
     end
 
+    rake_tasks do
+      load root.join("lib/tasks/rails_ai_build.rake")
+    end
+
     initializer "rails_ai_build.register_providers" do
       RailsAiBuild::Providers.register_defaults
     end